Bugfixed nullpointer-exceptions. App now runnable.

+++ src/se/chalmers/lidkopingsh/database/OrderDbStorage.java
@@ -152,12 +152,16 @@ public class OrderDbStorage {
 				order.getLastTimeUpdate());
 
 		insertCustomer(order.getCustomer());
-		for (Image i : order.getImages()) {
-			insertImage(i, order.getId());
+		if (order.getImages() != null) {
+			for (Image i : order.getImages()) {
+				insertImage(i, order.getId());
+			}
 		}
 
-		for (Product p : order.getProducts()) {
-			insertProduct(p, order.getId());
+		if (order.getProducts() != null) {
+			for (Product p : order.getProducts()) {
+				insertProduct(p, order.getId());
+			}
 		}
 
 		db.insert(OrderTable.TABLE_NAME, null, values);
@@ -299,23 +303,23 @@ public class OrderDbStorage {
 			db.delete(StoneTable.TABLE_NAME, StoneTable.COLUMN_NAME_PRODUCT_ID
 					+ EQUALS + QUESTION_MARK,
 					new String[] { String.valueOf(p.getId()) });
-
-			for (Task t : p.getTasks()) {
-				Cursor c = db//checks if there are any tasks with the given station if not the station is removed
+			if(p.getTasks() != null) {
+				for (Task t : p.getTasks()) {
+					Cursor c = db//checks if there are any tasks with the given station if not the station is removed
 						.query(TaskTable.TABLE_NAME, null,
 								TaskTable.COLUMN_NAME_STATION_ID + EQUALS
 										+ QUESTION_MARK, new String[] { String
 										.valueOf(t.getStation().getId()) },
 								null, null, null);
-				if (!c.moveToNext()) {
-					db.delete(StationTable.TABLE_NAME,
+					if (!c.moveToNext()) {
+						db.delete(StationTable.TABLE_NAME,
 							StationTable.COLUMN_NAME_STATION_ID + EQUALS
 									+ QUESTION_MARK, new String[] { String
 									.valueOf(t.getStation().getId()) });
-					stationIds.remove(t.getStation().getId());
+						stationIds.remove(t.getStation().getId());
+					}
 				}
 			}
-
 		}
 
 		db.delete(ImageTable.TABLE_NAME, ImageTable.COLUMN_NAME_ORDER_ID
